fixed range check bug, just introduced. Updated tests in build test

+++ tightdb_jni/src/com_tightdb_util.cpp
@@ -28,5 +28,5 @@ JNIEXPORT jint JNICALL Java_com_tightdb_util_nativeGetVersion(JNIEnv*, jclass)
 
 JNIEXPORT jlong JNICALL Java_com_tightdb_util_nativeGetInfiniteValue(JNIEnv*, jclass)
 {
-    return size_t(-1);
+    return -1;
 }
